Update to setting type

+++ main/src/main/java/net/citizensnpcs/npc/skin/SkinPacketTracker.java
@@ -129,7 +129,7 @@ public class SkinPacketTracker {
                     return;
 
                 updateNearbyViewers(entity.getNPC().data().get(NPC.Metadata.TRACKING_RANGE,
-                        Setting.NPC_SKIN_VIEW_DISTANCE.asDouble()));
+                        Setting.NPC_SKIN_VIEW_DISTANCE.asInt()));
             }
         }.runTaskLater(CitizensAPI.getPlugin(), 15);
     }
